REFACTOR: Move standard library of DSLs to own module.
