+++ packages/frontend/src/stdlib/index.ts
@@ -1,0 +1,10 @@
+/** Standard library of domain-specific logics.
+
+This module specifies the domain-specific logics that ship with Catcolab.
+Eventually it will be possible to extend Catcolab with user-specified logics but
+at present this is the complete set of logics available in the application.
+
+@module
+ */
+
+export { stdTheories } from "./theories";
