+++ packages/frontend/src/stdlib/svg_styles.module.css
